Add files via upload
Project Update 2 CoppeliaSim Scene. 

A 3D blob detector is used to locate the detectable object (green). The coordinates of the object are sent to the target of the inverse kinematics tip-target pair. The target is moved to the detected object and by inverse kinematics, the robotic arm is moved so the tip is also positioned at the detected object.